requirements.txt
asgiref==3.7.2
attrs==23.1.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.4
defusedxml==0.8.0rc1
Django==4.2.5
django-cors-headers==4.2.0
django-debug-toolbar==4.2.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.0
drf-spectacular==0.26.5
flake8==6.1.0
gunicorn==21.2.0
idna==3.4
inflection==0.5.1
isort==5.12.0
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
mccabe==0.7.0
numpy==1.26.0
oauthlib==3.2.2
openpyxl==3.1.2
packaging==23.1
pandas==2.1.1
psycopg2-binary==2.9.7
pycodestyle==2.11.0
pycparser==2.21
pyflakes==3.1.0
Pygments==2.16.1
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
redgreenunittest==0.1.1
referencing==0.30.2
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.10.3
six==1.16.0
social-auth-app-django==5.3.0
social-auth-core==4.4.2
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.5
XlsxWriter==3.1.6
